+++ tests/test_cli.py
@@ -30,24 +30,29 @@ class TestWithMocks:
         runner = ct.CliRunner(mix_stderr=False)
         with mock.patch("bio2zarr.vcf.explode_init", return_value=5) as mocked:
             result = runner.invoke(
-                cli.vcf2zarr, ["dexplode-init", "source", "dest", "5"], catch_exceptions=False
+                cli.vcf2zarr,
+                ["dexplode-init", "source", "dest", "5"],
+                catch_exceptions=False,
             )
             assert result.exit_code == 0
             assert len(result.stderr) == 0
             assert result.stdout == "5\n"
             mocked.assert_called_once_with(
-                ("source",),
                 "dest",
+                ("source",),
                 target_num_partitions=5,
                 worker_processes=1,
+                column_chunk_size=64,
                 show_progress=True,
             )
 
-    def test_vcf_dexplode_slice(self):
+    def test_vcf_dexplode_partition(self):
         runner = ct.CliRunner(mix_stderr=False)
-        with mock.patch("bio2zarr.vcf.explode_slice") as mocked:
+        with mock.patch("bio2zarr.vcf.explode_partition") as mocked:
             result = runner.invoke(
-                cli.vcf2zarr, ["dexplode-slice", "path", "1", "2"], catch_exceptions=False
+                cli.vcf2zarr,
+                ["dexplode-partition", "path", "1"],
+                catch_exceptions=False,
             )
             assert result.exit_code == 0
             assert len(result.stdout) == 0
@@ -55,9 +60,6 @@ class TestWithMocks:
             mocked.assert_called_once_with(
                 "path",
                 1,
-                2,
-                column_chunk_size=64,
-                worker_processes=1,
                 show_progress=True,
             )
 
@@ -172,6 +174,89 @@ class TestWithMocks:
             variants_chunk_size=None,
             show_progress=True,
         )
+
+
+class TestVcfEndToEnd:
+    data_path = "tests/data/vcf/sample.vcf.gz"
+
+    def test_dexplode(self, tmp_path):
+        icf_path = tmp_path / "icf"
+        runner = ct.CliRunner(mix_stderr=False)
+        result = runner.invoke(
+            cli.vcf2zarr,
+            f"dexplode-init {self.data_path} {icf_path} 5",
+            catch_exceptions=False,
+        )
+        assert result.exit_code == 0
+        assert result.stdout.strip() == "3"
+
+        for j in range(3):
+            result = runner.invoke(
+                cli.vcf2zarr,
+                f"dexplode-partition {icf_path} {j}",
+                catch_exceptions=False,
+            )
+            assert result.exit_code == 0
+        result = runner.invoke(
+            cli.vcf2zarr, f"dexplode-finalise {icf_path}", catch_exceptions=False
+        )
+        assert result.exit_code == 0
+
+        result = runner.invoke(
+            cli.vcf2zarr, f"inspect {icf_path}", catch_exceptions=False
+        )
+        assert result.exit_code == 0
+        # Arbitrary check
+        assert "CHROM" in result.stdout
+
+    def test_explode(self, tmp_path):
+        icf_path = tmp_path / "icf"
+        runner = ct.CliRunner(mix_stderr=False)
+        result = runner.invoke(
+            cli.vcf2zarr, f"explode {self.data_path} {icf_path}", catch_exceptions=False
+        )
+        assert result.exit_code == 0
+        result = runner.invoke(
+            cli.vcf2zarr, f"inspect {icf_path}", catch_exceptions=False
+        )
+        assert result.exit_code == 0
+        # Arbitrary check
+        assert "CHROM" in result.stdout
+
+    def test_encode(self, tmp_path):
+        icf_path = tmp_path / "icf"
+        zarr_path = tmp_path / "zarr"
+        runner = ct.CliRunner(mix_stderr=False)
+        result = runner.invoke(
+            cli.vcf2zarr, f"explode {self.data_path} {icf_path}", catch_exceptions=False
+        )
+        assert result.exit_code == 0
+        result = runner.invoke(
+            cli.vcf2zarr, f"encode {icf_path} {zarr_path}", catch_exceptions=False
+        )
+        assert result.exit_code == 0
+        result = runner.invoke(
+            cli.vcf2zarr, f"inspect {zarr_path}", catch_exceptions=False
+        )
+        assert result.exit_code == 0
+        # Arbitrary check
+        assert "variant_position" in result.stdout
+
+    def test_convert(self, tmp_path):
+        zarr_path = tmp_path / "zarr"
+        runner = ct.CliRunner(mix_stderr=False)
+        result = runner.invoke(
+            cli.vcf2zarr,
+            f"convert {self.data_path} {zarr_path}",
+            catch_exceptions=False,
+        )
+        assert result.exit_code == 0
+        result = runner.invoke(
+            cli.vcf2zarr, f"inspect {zarr_path}", catch_exceptions=False
+        )
+        assert result.exit_code == 0
+        # Arbitrary check
+        assert "variant_position" in result.stdout
 
 
 class TestVcfPartition: