+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-# 0.0.2 2024-03-xx
+# 0.0.2 2024-03-27
 
 - Merged 1D and 2D encode steps into one, and change rate reporting to bytes
 - Add --max-memory for encode
 - Change `chunk_width` to `samples_chunk_size` and `chunk_length` to `variants_chunk_size`
+- Various updates to the intermediate chunked format, with breaking change to version 0.2
+- Add distributed explode commands